change fetch to the routes component
